src/test/javascript/spec/app/entities/my-calendar/my-calendar-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import format from 'date-fns/format';
import parseISO from 'date-fns/parseISO';
import { DATE_TIME_LONG_FORMAT } from '@/shared/date/filters';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import MyCalendarUpdateComponent from '@/entities/my-calendar/my-calendar-update.vue';
import MyCalendarClass from '@/entities/my-calendar/my-calendar-update.component';
import MyCalendarService from '@/entities/my-calendar/my-calendar.service';

import TrainingService from '@/entities/training/training.service';

import UserService from '@/admin/user-management/user-management.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('MyCalendar Management Update Component', () => {
    let wrapper: Wrapper<MyCalendarClass>;
    let comp: MyCalendarClass;
    let myCalendarServiceStub: SinonStubbedInstance<MyCalendarService>;

    beforeEach(() => {
      myCalendarServiceStub = sinon.createStubInstance<MyCalendarService>(MyCalendarService);

      wrapper = shallowMount<MyCalendarClass>(MyCalendarUpdateComponent, {
        store,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          myCalendarService: () => myCalendarServiceStub,

          trainingService: () => new TrainingService(),

          userService: () => new UserService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('load', () => {
      it('Should convert date from string', () => {
        // GIVEN
        const date = new Date('2019-10-15T11:42:02Z');

        // WHEN
        const convertedDate = comp.convertDateTimeFromServer(date);

        // THEN
        expect(convertedDate).toEqual(format(date, DATE_TIME_LONG_FORMAT));
      });

      it('Should not convert date if date is not present', () => {
        expect(comp.convertDateTimeFromServer(null)).toBeNull();
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.myCalendar = entity;
        myCalendarServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(myCalendarServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.myCalendar = entity;
        myCalendarServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(myCalendarServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
